malawian_educational_ai_tutor/requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
fastapi
uvicorn[standard]
pydantic
pypdf
python-docx
textract
requests
